Track triples & lag per query host
This now uses the new headers to track
individual hosts.

For now we continue tracking the old metric!
(until we have a bit of data in the new location)
Then we can remove the old stuff and switch the graphs!

Note: we could make individual requests to each
host every min for this

Bug: T119993
Change-Id: Ie668c795d3c72ff39487b05a08e952869f59611b