Track global user enables & disables of beta features
This tracks how many global users have enabled / disabled
a betafeature.

If a user has a feature enabled on a single wiki they are
counted as using the feature.
If they then enabled it on a second wiki, this will not count
as another enable.
Only if a user removes the feature from all wikis will it count
as a disable.

Bug: T140226
Change-Id: Ib3df65f4e28bf2ffa06e21f1d4f89f13f79c6582