instanceof.php short sleep to avoid rate limit
This is probably an issue as the script now always
hits the same backend instance.

In theroy it shouldnt hit the rate limit as requests are
done in series, but I guess nginx doesnt figure out if
the rate limit has been met with that much accuracy.

It might be possible to remove this once T176875 is solved
but there is probably no point.

Bug: T176577
Change-Id: Ie743ccda8023bfedf5d61e1448f8b5180c4beb2a