Send maxlag & dispatch lag directly to graphite
Sending this data to statsd result in a delay and sometimes
interesting data making it to graphite depending on when statsd
buckets get emptied.

As we only check these once a minuite, and every minuite, and the
volume of data is very low we can talk directly to graphite to
avoid this.

Change-Id: If8f3335c910addd4a4447a7cd71013023f36d450