Change metric for TwoColConflict disables
We want to count all disables including both the beta feature and the
new preference.  The main motivation is that an automatic migration
from beta opt-out to new preference opt-out should not count against
adoption.

This will cause a discontinuity, and unfortunately there's no obvious
way to backfill the historical data.  The patch changes the metric
name, so we can mark the cutover in graphs and show both the old and
new totals.

Bug: T257577
Change-Id: I0f76dd64a656d99218326e2975e49d7628d77ed8