Optimize SPARQL query for ranks
With COUNT(DISTINCT ?s), Blazegraph actually scans all the rank triples,
which takes some ten to fifteen seconds; with COUNT(*), it can get the
count from some kind of metadata it has, within a few hundred
milliseconds. The result is the same: since an RDF graph cannot contain
duplicate triples, all bound ?s must be distinct anyways, and as the
query does not allow for ?s to be unbound, there can be no difference
between COUNT(?s) and COUNT(*) either.

Change-Id: I0326140a5be9e38eee5c63b1cb2ac2e56a466aaf