Send accurate timestamp with lexeme statistics
Previously, we let WikimediaGraphite::sendNow() calculate the current
time each time we called it, which I assume meant that in theory results
from the same query could be sent with slightly different timestamps
(though the timestamp only has a precision of one second, so I don’t
think this was likely to happen in practice). Let’s calculate one
timestamp per query and use that to send all the results.

The timestamp is taken just before we send the query to Blazegraph, not
afterwards. Blazegraph claims to be an ACID database [1] with isolated
transactions [2], which I assume means something like “REPEATABLE READ”
in SQL (rather than “READ COMMITTED” or weaker), so I think taking the
timestamp at the beginning of the query is more accurate than taking it
at the end: the query results ought to reflect the state of the database
when the query was started, even if the data changed while the query was
running.

[1]: https://dbdb.io/db/blazegraph
[2]: https://github.com/blazegraph/database/wiki/TxGuide#kinds-of-transactions

Change-Id: Ifb467e8586ff4e5d8d1d50551f81a12fb481db2d