Reduce duplicate code in lexeme statistics script
In the first version of the script, I thought that the different data
collection functions differed enough that extracting the common code
wasn’t a clear benefit, but with the addition of summarizing “other”
results (change I6ae6187ff6), this has arguably changed enough that it’s
now worth to extract one function to do the real work.

I tested locally that the script still returns mostly the same data
(except for minor differences due to running the queries at different
times, and sometimes getting a timeout).

Change-Id: I026fd8e7d1f185347569e71697e5fe1b8968a081