Track number of active items
An item is active if it has been edited in the past 30 days. For now, we
don’t care about the number of edits (>= 1), but in case that changes in
the future, we include the “1” in the metric name.

The code is mainly based on active_users.php and rolling_rc.php

The query is a bit more complex than a simple COUNT(DISTINCT rc_title),
but this is not only more flexible in case we want other thresholds
later, it’s even consistently a bit faster in my testing (17 instead of
20 seconds, plus or minus half a second).

Bug: T286903
Change-Id: Id8852aa39dccadfda24c7bf5bad1b4612bb8fb15