README: Clarify development process
IIUC, historically several development commits on the master branch
might be squashed into a single commit on the production branch, but I
think these days we mostly just cherry-pick each commit as it happens.
Document that a bit more explicitly, since it’s fairly different from a
lot of other repos we usually work with.

Change-Id: Ia389844964d8d0c400718dd1234eebf51fb629e3