Remove deprecated tech wish scripts
We decided to remove the scripts related to TechWishe projects.
Almost all of them are not relevant anymore and others are not
working as intended. For the relevant parts we might come up with
replacements, but that's for another time and codebase I guess.

Bug: T350411
Bug: T348644
Change-Id: Ibce227be0572d42c2433305276f53519167aec7a
(cherry picked from commit d318d6fdf751cae9a222b9e6e4aebe5a78cdf345)