Rewrite WikimediaDbSectionMapper::loadDbMap()
The old version broke due to change Ie326bab69f (commit ac74109fce) and
was pretty evil anyway. Instead of downloading the PHP config and
eval’ing it, let’s instead download all the dblist files (I think we
have to hardcode which ones exist, unfortunately, but hopefully that’ll
be relatively stable) and get the mapping from there. As a safety check,
also get all.dblist and throw an error if any wikis are listed there but
not in the section dblists (which probably means a new section was
added?) – hopefully we’ll see this relatively quickly.

Bug: T371706
Change-Id: Ieca0660a8d9f2e3859bdca1f0fa124602b599bef