Remove current date targetDate from Prometheus metrics
In these two scripts, the targetDate was just the result of

    date( DATE_ATOM )

i.e. the current date formatted in an ISO-8601-like format. I’m guessing
the date was included in the Graphite metrics to avoid different copies
of the metrics having slightly different timestamps, but in Prometheus
it seems that the label does more harm than good, so remove it.

Also fix some indentation in lexemes.php while we’re touching those
lines.

Bug: T389344
Change-Id: I0a80b559e48d91ddea65e5f00003349fbee3cadb