Send entity usage aspect with modifier combined to Prometheus
The Graphite code here effectively turned a database string like C.P1
into C_P1, thus ensuring that it would not be treated as a “hierarchy
level” in Graphite. To reduce cardinality in Prometheus, we should do
something similar on the Prometheus side: just track the entity usage
aspect including its modifier as one label, not split into two labels.
And as the . isn’t special in a Prometheus label value, we can simply
send the row value from the database directly as the label contents :)

Bug: T389344
Change-Id: I1659855128baf74f62cf19c365853d14855b8f7a