Remove last traces of Graphite, send data to Prometheus only
Remove WikimediaGraphite and all calls to it, and change some comments
from Graphite to Prometheus.

In active_users_by_namespace.php, also fix some statsd indentation.

In pages_by_namespace.php, do the same, and also remove $namespaceTotals
as it’s not sent to Prometheus. (I *vaguely* seem to recall us not
porting that metric over to Prometheus because it can be reconstructed
from the other one.)

In lexemes.php, rename the method and leave a comment about the
historical motivation for limiting the number of metrics.

In dumpDownloads.php, rename a variable that shouldn’t be named after
Graphite anymore.

In README.md, just remove the last section. I think the note about the
host names(?) being hard-coded was already outdated (see graphite_host),
and the metric names are now no longer hierarchically organized, so it
doesn’t really make sense to note what they start with.

Change-Id: I26d177176fb48f6311a75d04e95492b96ef543dd